+++ neurocaps/__init__.py
@@ -3,4 +3,4 @@ from . import analysis, extraction
 __all__ = ["analysis", "extraction"]
 
 # Version in single place
-__version__ = "0.21.1"
+__version__ = "0.21.2"