+++ CHANGELOG.md
@@ -13,6 +13,11 @@ noted in the changelog (e.g., new functions or parameters, changes in parameter 
 improvements/enhancements. All fixes and modifications are backwards compatible.
 - *.postN* : Consists of documentation changes or metadata-related updates, such as modifications to type hints.
 
+## [0.21.2] - 2025-01-14
+### 🐛 Fixes
+- Fixes warning about ignoring mandatory keys in `fd_threshold` and `dummy_scans`.
+- Also adds check to ensure that the "outlier_percentage" key is a float between 0 and 1.
+
 ## [0.21.1] - 2025-01-10
 ### 🐛 Fixes
 - Better type validation for `fd_threshold` and `dummy_scans`.